+++ crates/lang/tests/ui/trait_def/fail/message_selector_wildcard_selector_overlap.rs
@@ -1,0 +1,9 @@
+use ink_lang as ink;
+
+#[ink::trait_definition]
+pub trait TraitDefinition {
+    #[ink(message, selector = 1, selector = "_")]
+    fn message_1(&self);
+}
+
+fn main() {}